rewrite addchannel to REST
