+++ harmony-server/rest/setup.go
@@ -17,8 +17,10 @@ func apiV1(g echo.Group) {
 	r.POST("/login*", v1.WithRateLimit(v1.Login, 5 * time.Second, 1))
 	r.POST("/register*", v1.WithRateLimit(v1.Register, 10 * time.Minute, 10))
 	r.POST("/getguilds*", v1.WithRateLimit(v1.GetGuilds, 5 * time.Second, 3))
-	r.POST("/getchannels/:guildid", v1.WithRateLimit(v1.GetChannels, 500 * time.Second, 5))
+	r.POST("/getchannels/:guildid", v1.WithRateLimit(v1.GetChannels, 500 * time.Millisecond, 5))
 	r.POST("/avatarupdate*", v1.WithRateLimit(v1.AvatarUpdate, 3 * time.Second, 1))
 	r.POST("/updateguildpicture/:guildid/", v1.WithRateLimit(v1.UpdateGuildPicture, 3 * time.Second, 1))
-	r.POST("/message/:guildid/:channelid/*", v1.WithRateLimit(v1.Message, 500 * time.Millisecond, 20))
+
+	r.POST("/message/:guildid/:channelid/", v1.WithRateLimit(v1.Message, 500 * time.Millisecond, 20))
+	r.POST("/addchannel/:guildid", v1.WithRateLimit(v1.AddChannel, 1 * time.Second, 3))
 }
